% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{log of sum of exponentials function}
\usage{
log_summed_exps(vec)
}
\arguments{
\item{vec}{A numeric vector}
}
\value{
The sum of exponentials of the input vector
}
\description{
log of sum of exponentials function
}
\examples{
tmp <- seq(1:10) # create a vector
log_summed_exps(tmp) # call the function using this vector

}
